Advanced 16

+++ functions_nested_loops/104-fibonacci.c
@@ -7,16 +7,16 @@
 int main(void)
 {
 	int i;
-	unsigned long long n1 = 1;
-	unsigned long long n2 = 2;
-	unsigned long long fib;
+	unsigned long n1 = 1;
+	unsigned long n2 = 2;
+	unsigned long fib;
 
 	for (i = 1; i <=98; i++)
 	{
 		fib = n2 + n1;
 		n1 = n2;
 		n2 = fib;
-		printf("%llu, ", fib);
+		printf("%lu, ", fib);
 	}
 	printf("\n");
 	return (0);
